Added FindPathBehavior which finds the a path to a given target Entity. Added a secondary path generator that can generate a path to a list of points instead of a single point. Added FollowPathBehavior which replaces the MoveToLocationSystem.

+++ IsoECS/IsoECS/Entities/Entity.cs
@@ -17,12 +17,12 @@ namespace IsoECS.Entities
         public static uint ID_COUNTER = 0;
 
         public string UniqueID { get; set; }
-        public uint ID { get; private set; }
+        public int ID { get; private set; }
         public Dictionary<Type, Component> Components { get; private set; }
 
         public Entity()
         {
-            ID = ++ID_COUNTER;
+            ID = (int)++ID_COUNTER;
             Components = new Dictionary<Type, Component>();
         }
 
@@ -75,7 +75,7 @@ namespace IsoECS.Entities
 
         private void ResetID()
         {
-            ID = ++ID_COUNTER;
+            ID = (int)++ID_COUNTER;
         }
     }
 }
